# Improve performance of barrier()