--- a/test/python/circuit/test_circuit_operations.py
+++ b/test/python/circuit/test_circuit_operations.py
@@ -550,6 +550,17 @@ class TestCircuitOperations(QiskitTestCase):
 
         self.assertEqual(qc, expected)
 
+    def test_barrier_in_context(self):
+        """Test barrier statement in context, see gh-11345 """
+        qc = QuantumCircuit(2, 2)
+        qc.h(0)
+        with qc.if_test((qc.clbits[0], False)):
+            qc.h(0)
+            qc.barrier()
+
+        operation_names = [c.operation.name for c in qc]
+        assert 'barrier' not in operation_names
+
     def test_measure_active(self):
         """Test measure_active
         Applies measurements only to non-idle qubits. Creates a ClassicalRegister of size equal to