@eendebakpt 2024-01-03
@jakelishman Thanks for the pointer. How could I write a test for the control flow case? And what to test for exactly? I tried this, and it runs without errors.
```
from qiskit import QuantumCircuit, QuantumRegister, ClassicalRegister
from qiskit.circuit import IfElseOp

true_body = QuantumCircuit(2)
true_body.x(1)
true_body.barrier() # barrier inside conditional

qc = QuantumCircuit(2,2)
qc.h(0)
qc.measure(0, 0)

op = IfElseOp((qc.clbits[0], True), true_body, None)
qc.append(op, qargs=[0, 1], cargs=[])

qc.draw()
```