@jakelishman 2024-01-03
Right - `QuantumCircuit.append` (no underscore) respects control-flow builder scopes already, but `QuantumCircuit._append` is the internal worker of the `QuantumCircuit` innermost scope, so calling it (deliberately) skips any active builder scope.  For example, using commit 35ec9ee as the current head of this PR:
```python
﻿﻿from qiskit import QuantumCircuit

qc = QuantumCircuit(2, 2)
qc.h(0)
qc.cx(0, 1)
qc.measure([0, 1], [0, 1])
with qc.if_test((qc.clbits[0], False)):
    qc.h(0)
    qc.barrier()
    qc.h(0)
    qc.barrier()
    qc.h(0)
qc.draw()
```
```
     ┌───┐     ┌─┐    ░  ░   ┌──────  ┌───┐┌───┐┌───┐ ───────┐
q_0: ┤ H ├──■──┤M├────░──░───┤ If-0  ─┤ H ├┤ H ├┤ H ├  End-0 ├─
     └───┘┌─┴─┐└╥┘┌─┐ ░  ░   └──╥───  └───┘└───┘└───┘ ───────┘
q_1: ─────┤ X ├─╫─┤M├─░──░──────╫──────────────────────────────
          └───┘ ║ └╥┘ ░  ░ ┌────╨────┐
c: 2/═══════════╩══╩═══════╡ c_0=0x0 ╞═════════════════════════
                0  1       └─────────┘
```
where the barriers have migrated outside the "if".  The result should look like if I'd manually constructed all the scopes explicitly:
```python

from qiskit import QuantumCircuit

qc = QuantumCircuit(2, 2)
qc.h(0)
qc.cx(0, 1)
qc.measure([0, 1], [0, 1])

true_body = QuantumCircuit(qc.qubits, [qc.clbits[0]])
true_body.h(0)
true_body.barrier()
true_body.h(0)
true_body.barrier()
true_body.h(0)

qc.if_test((qc.clbits[0], False), true_body, qc.qubits, [qc.clbits[0]])
qc.draw()
```
```
     ┌───┐     ┌─┐     ┌──────  ┌───┐ ░ ┌───┐ ░ ┌───┐ ───────┐
q_0: ┤ H ├──■──┤M├─────┤       ─┤ H ├─░─┤ H ├─░─┤ H ├        ├─
     └───┘┌─┴─┐└╥┘┌─┐  │ If-0   └───┘ ░ └───┘ ░ └───┘  End-0 │
q_1: ─────┤ X ├─╫─┤M├──┤       ───────░───────░──────        ├─
          └───┘ ║ └╥┘  └──╥───        ░       ░       ───────┘
                ║  ║ ┌────╨────┐
c: 2/═══════════╩══╩═╡ c_0=0x0 ╞═══════════════════════════════
                0  1 └─────────┘
```