@eendebakpt 2024-01-15
Updated benchmark:
```
def g():
    qc = QuantumCircuit(2, 2)
    for ii in range(5):
        qc.barrier(qc.qubits)
```
Output of `%timeit g()` for main and this PR:
```
main: 259 µs ± 18.7 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
pr:149 µs ± 8.41 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```